refactor(case_search): move distance units out of es ap
